requirements.txt
autopep8==1.2.2
cookies==2.2.1
flake8==2.5.4
mccabe==0.4.0
nose==1.3.7
pep8==1.7.0
pyflakes==1.0.0
requests==2.9.1
responses==0.5.1
six==1.10.0
python-daemon==2.1.1
